+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Plugin `excalidraw`: Integrate Excalidraw in SysReptor
 * PDF viewer: fix single-letter shortcuts applied when typing in searchbar input
 * Fix scroll sync to wrong position on HTML blocks starting with `<br>` tags
+* UI: add blockquote button to markdown editor toolbar
 
 
 ## v2025.56 - 2025-06-17